* Improved a bit

+++ visualiser/Test.cpp
@@ -12,9 +12,9 @@ int main() {
 
 	float c;
 
-	for (unsigned int w = 0; w < width; w++) {
-		c = 2.0f * ((float)w) / ((float)width) - 1.0f;
-		for (unsigned int h = 0; h < height; h++) {
+	for (int w = 0; w < width; w++) {
+		c = (2.0f * w / width) - 1.0f;
+		for (int h = 0; h < height; h++) {
 			altitudes[w*height + h] = c;
 		}
 	}
